requirements.txt
python-dotenv~=1.0.1
discord~=2.3.2c